freezing

+++ src/main/java/ai/ilikeplaces/logic/cdn/CDNAlbum.java
@@ -49,9 +49,11 @@ public class CDNAlbum extends CDN implements CDNAlbumLocal {
     private static final String ALBUM_PHOTO_UPLOAD_FAILED_DUE_TO_RENAMING_ISSUES = "Album Photo Upload Failed Due To Renaming Issues!";
     private static final String ALBUM_PHOTO_UPLOAD_SUCCESSFUL = "Album Photo Upload Successful.";
     private static final String UPLOADING_IMAGE = "Uploading Image";
+    private static final String UPLOADING_IMAGE_THUMB = "Uploading Image Thumbnail";
     private static final String SAVING_SCALED_IMAGE = "Saving Scaled Image";
     private static final String SCALING_IMAGE = "Scaling Image";
     private static final String LOADING_IMAGE_AS_BUFFERED_IMAGE = "Loading Image As Buffered Image";
+    public static final String THUMBNAIL = "th_";
 
     public static CDNAlbumLocal getAlbumPhotoCDNLocal() {
         isOK();
@@ -112,7 +114,21 @@ public class CDNAlbum extends CDN implements CDNAlbumLocal {
                         final String cdnFileName = newFile.getName();
                         sl.appendToLogMSG(UPLOADING_IMAGE);
                         final boolean uploaded = client.storeObjectAs(CONTAINER, newFile, FilesConstants.getMimetype(userFileExtension), cdnFileName);
-                        if (uploaded) {
+
+                        sl.appendToLogMSG(LOADING_IMAGE_AS_BUFFERED_IMAGE);
+                        BufferedImage bi = loadImage(newFile);
+
+                        sl.appendToLogMSG(SCALING_IMAGE);
+                        bi = scaleImage(bi, 190); //Reducing size of image to blueprintcss span-5 just to save bandwidth for the user.
+
+                        sl.appendToLogMSG(SAVING_SCALED_IMAGE);
+                        saveImage(bi, newFile);
+
+                        final String cdnThumbFileName = THUMBNAIL + newFile.getName();
+                        sl.appendToLogMSG(UPLOADING_IMAGE_THUMB);
+                        final boolean uploadedThumb = client.storeObjectAs(CONTAINER, newFile, FilesConstants.getMimetype(userFileExtension), cdnThumbFileName);
+
+                        if (uploaded && uploadedThumb) {
                             final boolean deleted = newFile.delete();
                             if (deleted) {
                                 final Return<Album> dbr = DB.getHumanCrudPrivateEventLocal(true).uPrivateEventAddEntryToAlbum(humanId, Long.parseLong((String) parameterMap.get(ALBUM_PIVATE_EVENT_ID)), new Obj<String>(cdnFileName).getSelfAsValid());
